% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dice.major}
\alias{dice.major}
\title{Dice Major}
\format{
An object of class GMT; see \code{read.gmt} from qusage package.
}
\usage{
data(dice.major)
}
\description{
Dice Major
}
\references{
Schmiedel et al. (2018) Immunity 175:1701-1715
(\doi{10.1016/j.cell.2018.10.022})
}
\keyword{datasets}
