% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC.scRNAseq}
\alias{PBMC.scRNAseq}
\title{PBMC scRNAseq modules}
\format{
An object of class GMT; see \code{read.gmt} from qusage package.
}
\usage{
data(PBMC-scRNAseq)
}
\description{
PBMC scRNAseq modules
}
\keyword{datasets}
