% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki}
\alias{wiki}
\title{Wikipathways}
\format{
An object of class GMT; see \code{read.gmt} from qusage package.
}
\usage{
data(wiki)
}
\description{
Wikipathways
}
\references{
Pico et al. (2008) PLOS Biology 
(\doi{10.1371/journal.pbio.0060184})
}
\keyword{datasets}
