% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC.scRNAseq}
\alias{PBMC.scRNAseq}
\title{PBMC scRNAseq modules}
\format{
An object of class GMT; see \code{read.gmt} from qusage package.
}
\usage{
data(PBMC.scRNAseq)
}
\description{
PBMC scRNAseq modules
}
\references{
Márquez et al., 2020; Nehar-Belaid et al., 2020
\doi{10.1038/s41590-020-0743-0}
\doi{10.1038/s41467-020-14396-9}
}
\keyword{datasets}
